closes #16 - precompiled-builds
