+++ package.json
@@ -1,14 +1,22 @@
 {
   "name": "@pderas/vue2-table",
-  "version": "1.3.2",
+  "version": "2.0.0",
   "description": "A simple vue component for rendering sortable and searchable tables.",
-  "main": "index.js",
+  "main": "dist/vue-table.umd.min.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "lint": "eslint src --ext .js,.vue",
+    "lint:fix": "eslint src --ext .js,.vue --fix",
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "bundle:project": "vue-cli-service build --target lib --name vue-table ./src/index.js",
+    "bundle": "npm run lint && npm run bundle:project"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/PDERAS/vue2-table.git"
+    "url": "git@github.com:PDERAS/vue2-table.git"
+  },
+  "publishConfig": {
+    "registry": "https://npm.pkg.github.com/"
   },
   "keywords": [
     "vue",
@@ -24,9 +32,25 @@
   "author": "David Buss",
   "license": "MIT",
   "peerDependencies": {
-    "vue": "2.x"
+    "vue": "2.x",
+    "axios": "^0.19"
   },
-  "devDependancies": {
-      "axios": "^0.17"
+  "files": [
+    "dist/*",
+    "src/*"
+  ],
+  "devDependencies": {
+    "@vue/babel-preset-app": "^3.12.0",
+    "@vue/cli-plugin-babel": "^3.12.0",
+    "@vue/cli-plugin-eslint": "^3.12.0",
+    "@vue/cli-service": "^3.12.0",
+    "eslint": "^6.5.1",
+    "eslint-config-prettier": "^6.4.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-vue": "^5.2.3",
+    "fibers": "^4.0.1",
+    "sass": "^1.23.0",
+    "sass-loader": "^8.0.0",
+    "vue-template-compiler": "^2.6.10"
   }
 }
