+++ vue.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    lintOnSave: process.env.NODE_ENV !== 'production',
+    css: {
+      loaderOptions: {
+        sass: {
+          implementation: require('sass')
+        }
+      }
+    }
+  }
