+++ yarn-error.log
@@ -1,0 +1,80 @@
+Arguments: 
+  /Users/reed/.nvm/versions/node/v12.10.0/bin/node /Users/reed/.yarn/bin/yarn.js add -D vue-cli-plugin-babel
+
+PATH: 
+  /Users/reed/code/Plugins/vue2-table/node_modules/.bin:/Users/reed/code/Plugins/node_modules/.bin:/Users/reed/code/node_modules/.bin:/Users/reed/node_modules/.bin:/Users/node_modules/.bin:/node_modules/.bin:/Users/reed/.nvm/versions/node/v12.10.0/bin:/Users/reed/google-cloud-sdk/bin:/Users/reed/.yarn/bin:/Users/reed/.config/yarn/global/node_modules/.bin:/opt/local/bin:/opt/local/sbin:/Users/reed/.fastlane/bin:/Users/reed/.nvm/versions/node/v12.10.0/bin:/Library/Frameworks/Python.framework/Versions/3.6/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Users/reed/google-cloud-sdk/bin:/Users/reed/.yarn/bin:/Users/reed/.config/yarn/global/node_modules/.bin:/opt/local/bin:/opt/local/sbin:/Users/reed/.fastlane/bin:/Users/reed/.nvm/versions/node/v12.10.0/bin:/Library/Frameworks/Python.framework/Versions/3.6/bin:/Users/reed/.composer/vendor/bin:/Users/reed/Library/Android/sdk/tools:/Users/reed/Library/Android/sdk/tools/bin:/Users/reed/Library/Android/sdk/platform-tools:/Users/reed/.composer/vendor/bin:/Users/reed/Library/Android/sdk/tools:/Users/reed/Library/Android/sdk/tools/bin:/Users/reed/Library/Android/sdk/platform-tools
+
+Yarn version: 
+  1.16.0
+
+Node version: 
+  12.10.0
+
+Platform: 
+  darwin x64
+
+Trace: 
+  Error: https://registry.yarnpkg.com/vue-cli-plugin-babel: Not found
+      at Request.params.callback [as _callback] (/Users/reed/.yarn/lib/cli.js:66091:18)
+      at Request.self.callback (/Users/reed/.yarn/lib/cli.js:129590:22)
+      at Request.emit (events.js:209:13)
+      at Request.<anonymous> (/Users/reed/.yarn/lib/cli.js:130562:10)
+      at Request.emit (events.js:209:13)
+      at IncomingMessage.<anonymous> (/Users/reed/.yarn/lib/cli.js:130484:12)
+      at Object.onceWrapper (events.js:298:28)
+      at IncomingMessage.emit (events.js:214:15)
+      at endReadableNT (_stream_readable.js:1178:12)
+      at processTicksAndRejections (internal/process/task_queues.js:80:21)
+
+npm manifest: 
+  {
+    "name": "@pderas/vue2-table",
+    "version": "2.0.0",
+    "description": "A simple vue component for rendering sortable and searchable tables.",
+    "main": "dist/vue-table.umd.min.js",
+    "scripts": {
+      "test": "echo \"Error: no test specified\" && exit 1",
+      "serve": "vue-cli-service serve",
+      "build": "vue-cli-service build",
+      "bundle": "vue-cli-service build --target lib --name vue-table ./src/index.js"
+    },
+    "repository": {
+      "type": "git",
+      "url": "git@github.com:PDERAS/vue2-table.git"
+    },
+    "publishConfig": {
+      "registry": "https://npm.pkg.github.com/"
+    },
+    "keywords": [
+      "vue",
+      "vue2",
+      "tables",
+      "datatables",
+      "sorting",
+      "data",
+      "pderas",
+      "vue-table",
+      "vue2-table"
+    ],
+    "author": "David Buss",
+    "license": "MIT",
+    "peerDependencies": {
+      "vue": "2.x",
+      "axios": "^0.19"
+    },
+    "files": [
+      "dist/*"
+    ],
+    "devDependencies": {
+      "@vue/cli-service": "^3.12.0",
+      "sass": "^1.23.0",
+      "sass-loader": "^8.0.0",
+      "vue-template-compiler": "^2.6.10"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  No lockfile
